# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AFRICA CDC KnowledgeHub API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/knowledge_hub",
            "description": "Current deployment (set APP_URL in .env)"
        }
    ],
    "paths": {
...
        "/api/push-notifications": {
            "get": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Get list of push notifications",
                "description": "Returns list of push notifications",
                "operationId": "getPushNotificationsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Store new push notification",
                "description": "Returns push notification data",
                "operationId": "storePushNotification",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/push-notifications/{id}": {
            "get": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Get push notification information",
                "description": "Returns push notification data",
                "operationId": "getPushNotificationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Push notification not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Update existing push notification",
                "description": "Returns updated push notification data",
                "operationId": "updatePushNotification",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Push notification not found"
                    }
                }
            }
        },
        "/api/push-notifications/user": {
            "get": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Get push notifications by user",
                "description": "Returns list of push notifications for a specific user",
                "operationId": "getPushNotificationsByUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/push-notifications/mark-as-read": {
            "post": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Mark notifications as read",
                "description": "Marks specified notifications as read",
                "operationId": "markNotificationsAsRead",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notifications marked as read successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/push-notifications/unread-count": {
            "get": {
                "tags": [
                    "PushNotifications"
                ],
                "summary": "Get count of unread notifications",
                "description": "Returns the count of unread notifications for the authenticated user",
                "operationId": "getUnreadNotificationsCount",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}